Dodanie autoryzacji użytkownika

+++ src/main/java/pl/piwowarski/facebookly/service/entityService/CommentService.java
@@ -4,22 +4,31 @@ import lombok.AllArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.piwowarski.facebookly.exception.AccessDeniedException;
 import pl.piwowarski.facebookly.exception.CommentContentIsNullException;
 import pl.piwowarski.facebookly.exception.NoCommentWithSuchIdException;
+import pl.piwowarski.facebookly.exception.NoPostWithSuchIdException;
 import pl.piwowarski.facebookly.model.dto.CommentDto;
 import pl.piwowarski.facebookly.model.entity.Comment;
+import pl.piwowarski.facebookly.model.entity.Post;
+import pl.piwowarski.facebookly.model.enums.Role;
 import pl.piwowarski.facebookly.repository.CommentRepository;
+import pl.piwowarski.facebookly.repository.PostRepository;
 import pl.piwowarski.facebookly.service.mapper.map.impl.CommentMapper;
 import pl.piwowarski.facebookly.service.mapper.reverseMap.impl.CommentReverseMapper;
 
 import java.util.Comparator;
 import java.util.List;
 
+import static pl.piwowarski.facebookly.model.enums.Role.ADMIN;
+import static pl.piwowarski.facebookly.model.enums.Role.USER;
+
 @Service
 @AllArgsConstructor
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
     private final CommentMapper commentMapper;
     private final CommentReverseMapper commentReverseMapper;
 
@@ -51,7 +60,17 @@ public class CommentService {
         return commentMapper.map(savedComment);
     }
 
-    public void deleteById(Long commentId) {
+    public void deleteById(Long commentId, Long userId, Role role) {
+        Comment comment = commentRepository
+                .findById(commentId)
+                .orElseThrow(() -> new NoCommentWithSuchIdException(NoCommentWithSuchIdException.MESSAGE));
+        if(!comment.getUser().getId().equals(userId) && role != ADMIN){
+            throw new AccessDeniedException(AccessDeniedException.MESSAGE);
+        }
+        commentRepository.deleteById(commentId);
+    }
+
+    public void deleteById(Long commentId){
         commentRepository.deleteById(commentId);
     }
 
@@ -90,8 +109,14 @@ public class CommentService {
         comments.forEach(comment -> deleteById(comment.getId()));
     }
 
-    public void deleteAllPostComments(Long postId) {
-        List<Comment> comments = commentRepository.findAllByPostId(postId);
+    public void deleteAllPostComments(Long postId, Long userId, Role role) {
+        Post post = postRepository
+                .findById(postId)
+                .orElseThrow(() -> new NoPostWithSuchIdException(NoPostWithSuchIdException.MESSAGE));
+        if(!post.getUser().getId().equals(userId) && role == USER){
+            throw new AccessDeniedException(AccessDeniedException.MESSAGE);
+        }
+        List<Comment> comments = post.getComments();
         commentRepository.deleteAll(comments);
     }
 }
